Updates for doc branch.
