+++ Recipes/handle_error/handle_error_activities.rb
