+++ Recipes/pick_first_branch/search_activities.rb
