+++ Samples/hello_world_tutorial/helloworld_activity.rb
@@ -1,16 +1,15 @@
 require_relative 'helloworld_utils'
 
-# HelloWorldActivity class defines a set of activities for the HelloWorld sample.
+# Defines a set of activities for the HelloWorld sample.
 class HelloWorldActivity
   extend AWS::Flow::Activities
 
-  # The activity method is used to define activities. It accepts a list of names
-  # of activities and a block specifying registration options for those
-  # activities
+  # define an activity with the #activity method.
   activity :say_hello do
     {
       version: HelloWorldUtils::ACTIVITY_VERSION,
       default_task_list: HelloWorldUtils::ACTIVITY_TASKLIST,
+      # timeout values are in seconds.
       default_task_schedule_to_start_timeout: 30,
       default_task_start_to_close_timeout: 30
     }
@@ -23,4 +22,4 @@ class HelloWorldActivity
 end
 
 # Start an ActivityWorker to work on the HelloWorldActivity tasks
-HelloWorldUtils.new.activity_worker.start if $0 == __FILE__
+HelloWorldUtils.new.activity_worker(HelloWorldActivity).start if $0 == __FILE__
