+++ Samples/hello_world_tutorial/helloworld_workflow_starter.rb
@@ -2,4 +2,4 @@ require_relative 'helloworld_utils'
 require_relative 'helloworld_workflow'
 require_relative 'helloworld_activity'
 
-HelloWorldUtils.new.workflow_client.start_execution("AWS Flow Framework")
+HelloWorldUtils.new.workflow_client(HelloWorldWorkflow).start_execution("AWS Flow Framework")
